feature loading correction
