+++ app/src/main/java/caatsoft/studio/com/gamecatalog/auth/LoginActivity.kt
@@ -13,6 +13,8 @@ import androidx.core.content.ContextCompat
 import caatsoft.studio.com.gamecatalog.ui.MainActivity
 import caatsoft.studio.com.gamecatalog.R
 import caatsoft.studio.com.gamecatalog.databinding.ActivityLoginBinding
+import caatsoft.studio.com.gamecatalog.dismissDialog
+import caatsoft.studio.com.gamecatalog.startLoading
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseAuth.AuthStateListener
 
@@ -46,22 +48,22 @@ class LoginActivity : AppCompatActivity() {
         })
 
         activityLoginBinding.loginButton.setOnClickListener {
-            startLoading()
+            alertDialog = startLoading()
             val email = activityLoginBinding.emailEditView.text.toString()
             val password = activityLoginBinding.passwordEditView.text.toString()
             if (email.isEmpty() || email == null ||
                     password.isEmpty() || password == null) {
-                dismissDialog()
+                dismissDialog(alertDialog)
                 Toast.makeText(this@LoginActivity, getString(R.string.please_fill_in_what_is_missing), Toast.LENGTH_SHORT).show()
             } else {
                 mAuth.signInWithEmailAndPassword(email, password).addOnCompleteListener(this@LoginActivity) { task ->
                     if (!task.isSuccessful) {
-                        dismissDialog()
+                        dismissDialog(alertDialog)
                         Toast.makeText(this@LoginActivity, getString(R.string.there_was_an_authentication_error_please_try_to_register_again_or_try_to_login), Toast.LENGTH_SHORT).show()
                     } else{
                         val intent = Intent(this@LoginActivity, MainActivity::class.java)
                         startActivity(intent)
-                        dismissDialog()
+                        dismissDialog(alertDialog)
                         finish()
                     }
                 }
@@ -102,18 +104,5 @@ class LoginActivity : AppCompatActivity() {
 
     companion object {
         private const val REQUEST_CODE = 1
-    }
-
-    fun startLoading() {
-        val builder = AlertDialog.Builder(this)
-        val inflater = this.layoutInflater
-        builder.setView(inflater.inflate(R.layout.popup_loading, null))
-        builder.setCancelable(true)
-        alertDialog = builder.create()
-        alertDialog.show()
-    }
-
-    fun dismissDialog() {
-        alertDialog.dismiss()
     }
 }
