+++ app/src/main/java/caatsoft/studio/com/gamecatalog/auth/RegisterActivity.kt
@@ -16,7 +16,9 @@ import androidx.core.content.ContextCompat
 import caatsoft.studio.com.gamecatalog.ui.MainActivity
 import caatsoft.studio.com.gamecatalog.R
 import caatsoft.studio.com.gamecatalog.databinding.ActivityRegisterBinding
+import caatsoft.studio.com.gamecatalog.dismissDialog
 import caatsoft.studio.com.gamecatalog.network.model.DataUser
+import caatsoft.studio.com.gamecatalog.startLoading
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.FirebaseDatabase
@@ -49,6 +51,7 @@ class RegisterActivity : AppCompatActivity() {
         applyVerifyPermissions()
         activityRegisterBinding.register.setOnClickListener(View.OnClickListener {
             startLoading()
+            alertDialog =  startLoading()
             saveImage()
         })
         activityRegisterBinding.profileImageView.setOnClickListener(View.OnClickListener {
@@ -72,7 +75,7 @@ class RegisterActivity : AppCompatActivity() {
         if (name == null || name.isEmpty() || email == null
                 || email.isEmpty() || password == null || password.isEmpty()|| photourl == null || photourl.isEmpty()) {
             Toast.makeText(this, getString(R.string.all_must_be_filled_in), Toast.LENGTH_SHORT).show()
-            dismissDialog()
+            dismissDialog(alertDialog)
         } else{
             FirebaseAuth.getInstance().createUserWithEmailAndPassword(email, password)
                     .addOnCompleteListener { task ->
@@ -81,7 +84,7 @@ class RegisterActivity : AppCompatActivity() {
                             val intent = Intent(this@RegisterActivity, MainActivity::class.java)
                             intent.flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
                             startActivity(intent)
-                            dismissDialog()
+                            dismissDialog(alertDialog)
                             finish()
                         } else {
                             Toast.makeText(this@RegisterActivity, getString(R.string.enter_a_valid_email_address), Toast.LENGTH_LONG).show()
@@ -152,7 +155,7 @@ class RegisterActivity : AppCompatActivity() {
                 Toast.makeText(baseContext, getString(R.string.erro), Toast.LENGTH_LONG).show()
             }
         } else{
-            dismissDialog()
+            dismissDialog(alertDialog)
             Toast.makeText(this, getString(R.string.all_must_be_filled_in), Toast.LENGTH_SHORT).show()
         }
     }
@@ -196,18 +199,5 @@ class RegisterActivity : AppCompatActivity() {
         private const val PATH_LINK2 = ".jpg"
         private const val PATH_USER = "User"
         private const val TYPE_DATA = "image/*"
-    }
-
-    fun startLoading() {
-        val builder = AlertDialog.Builder(this)
-        val inflater = this.layoutInflater
-        builder.setView(inflater.inflate(R.layout.popup_loading, null))
-        builder.setCancelable(true)
-        alertDialog = builder.create()
-        alertDialog.show()
-    }
-
-    fun dismissDialog() {
-        alertDialog.dismiss()
     }
 }
